StoreKPLite/bot_tg/requirements.txt
aiogram>=3.4
httpx>=0.25
python-dotenv>=1.0
